WIP create flow
