+++ apps/waterfall/waterfall/src/app/dashboard/dashboard.module.ts
@@ -50,7 +50,7 @@ const routes: Routes = [{
         {
           path: 'create',
           data: { createMode: true },
-          loadChildren: () => import('@blockframes/waterfall/dashboard/edit/edit.module').then(m => m.WaterfallEditFormModule),
+          loadChildren: () => import('@blockframes/waterfall/dashboard/edit-title/edit.module').then(m => m.WaterfallEditTitleModule),
         },
         {
           path: ':movieId',
@@ -62,9 +62,14 @@ const routes: Routes = [{
               data: { redirect: '/c/o/dashboard/home' }
             },
             {
+              path: 'init',
+              data: { createMode: true },
+              loadChildren: () => import('@blockframes/waterfall/dashboard/edit-waterfall/edit.module').then(m => m.WaterfallEditFormModule),
+            },
+            {
               path: 'edit',
               data: { createMode: false },
-              loadChildren: () => import('@blockframes/waterfall/dashboard/edit/edit.module').then(m => m.WaterfallEditFormModule),
+              loadChildren: () => import('@blockframes/waterfall/dashboard/edit-title/edit.module').then(m => m.WaterfallEditTitleModule),
             },
             {
               path: 'right-holders-management',
