+++ apps/waterfall/waterfall/src/app/dashboard/title/view/view.component.html
@@ -1,8 +1,2 @@
-<waterfall-dashboard-shell [routes]="navLinks" editRoute="./edit">
-  <waterfall-cta fxLayout="column" fxLayout.lt-md="row" fxLayoutAlign="space-between end" fxFlex>
-    <a mat-button color="primary" [routerLink]="['waterfall']" class="main-dashboard-cta">
-      <mat-icon svgIcon="water_drop"></mat-icon>
-      <span>Waterfall Management</span>
-    </a>
-  </waterfall-cta>
-</waterfall-dashboard-shell>
+
+<waterfall-dashboard-shell [routes]="navLinks" editRoute="./edit"></waterfall-dashboard-shell>
