+++ apps/waterfall/waterfall/src/app/dashboard/title/view/view.component.ts
@@ -13,19 +13,21 @@ export class TitleViewComponent {
   navLinks: RouteDescription[] = [
     {
       path: 'dashboard',
-      label: 'Dashboard'
+      label: 'Dashboard',
     },
     {
       path: 'statements',
-      label: 'Statements'
+      label: 'Statements',
+      shouldHide: true,
     },
     {
       path: 'documents',
-      label: 'Documents'
+      label: 'Documents',
     },
     {
       path: 'waterfall',
       label: 'Waterfall',
+      shouldHide: true,
     },
     {
       path: 'right-holders',
@@ -34,6 +36,7 @@ export class TitleViewComponent {
     {
       path: 'sales',
       label: 'World Sales',
+      shouldHide: true,
     },
   ];
