+++ libs/waterfall/src/lib/components/contracts-form/contracts-form.component.html
@@ -2,10 +2,6 @@
 
   <article class="header">
     <h2>Contracts</h2>
-    <button mat-button color="primary" (click)="skip.emit()">
-      <span>Skip and do it later</span>
-      <mat-icon svgIcon="arrow_forward"></mat-icon>
-    </button>
   </article>
 
   <p>Please provide us with all the contracts, so that we can create a waterfall.</p>
