+++ libs/waterfall/src/lib/components/right-holder-form/right-holder-form.component.scss
@@ -32,4 +32,11 @@ section {
     }
   }
 
+  article.no-rightholders {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+  }
+
 }
