+++ libs/waterfall/src/lib/dashboard/edit-title/edit.component.html
@@ -1,0 +1,29 @@
+
+<ng-container *ngIf="loading$ | async; else ready">
+  <article>
+    <logo-spinner></logo-spinner>
+  </article>
+</ng-container>
+<ng-template #ready>
+  <a mat-icon-button routerLink="..">
+    <mat-icon svgIcon="arrow_back"></mat-icon>
+  </a>
+  <h1>{{ createMode ? 'Create' : 'Edit' }} Movie</h1>
+    <waterfall-title-form
+      [movieId]="movieId"
+      [movieForm]="movieForm"
+      [createMode]="createMode"
+    ></waterfall-title-form>
+
+    <section>
+      <ng-container *ngIf="updating$ | async; else editInfo">
+        <logo-spinner></logo-spinner>
+      </ng-container>
+      <ng-template #editInfo>
+        <button mat-flat-button color="primary" (click)="update()" [disabled]="invalidMovie$ | async">
+          <span>{{ createMode ? 'Create' : 'Edit' }} Movie</span>
+          <mat-icon svgIcon="arrow_forward"></mat-icon>
+        </button>
+      </ng-template>
+    </section>
+</ng-template>
