+++ libs/waterfall/src/lib/dashboard/edit-waterfall/edit.component.scss
@@ -1,0 +1,40 @@
+
+:host {
+  display: block;
+  padding: 48px 24px;
+}
+
+article {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin-top: 32px;
+
+  a.save {
+    margin-top: 24px;
+  }
+
+
+}
+
+section {
+  gap: 12px;
+  display: flex;
+  justify-content: center;
+
+  a.surface {
+    cursor: pointer;
+    margin-top: 42px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    width: 330px;
+    
+    img {
+      width: 100px;
+      align-self: center;
+    }
+  }
+}
