--- libs/waterfall/src/lib/dashboard/edit/edit.component.html
@@ -1,83 +1,0 @@
-
-<ng-container *ngIf="loading$ | async; else ready">
-  <article>
-    <logo-spinner></logo-spinner>
-  </article>
-</ng-container>
-<ng-template #ready>
-  <a mat-icon-button routerLink="..">
-    <mat-icon svgIcon="arrow_back"></mat-icon>
-  </a>
-  <h1>{{ createMode ? 'Create' : 'Edit' }} Movie</h1>
-
-  <mat-stepper #stepper>
-    <!-- override default stepper icons for the "edit" & "done" steps -->
-    <ng-template matStepperIcon="edit" let-index="index">
-      {{index + 1}}
-    </ng-template>
-    <ng-template matStepperIcon="done" let-index="index">
-      {{index + 1}}
-    </ng-template>
-
-    <mat-step [completed]="true">
-      <ng-template matStepLabel>Movie Information</ng-template>
-
-      <waterfall-title-form
-        [movieId]="movieId"
-        [movieForm]="movieForm"
-        [createMode]="createMode"
-      ></waterfall-title-form>
-
-      <section>
-        <ng-container *ngIf="updating$ | async; else editInfo">
-          <logo-spinner></logo-spinner>
-        </ng-container>
-        <ng-template #editInfo>
-          <button mat-flat-button color="primary" (click)="update(true)" [disabled]="invalidMovie$ | async">
-            <span>{{ createMode ? 'Create' : 'Edit' }} Movie</span>
-            <mat-icon svgIcon="arrow_forward"></mat-icon>
-          </button>
-        </ng-template>
-      </section>
-    </mat-step>
-
-    <mat-step>
-      <ng-template matStepLabel>Right Holders Information</ng-template>
-      <waterfall-right-holder-form
-        [rightholdersForm]="rightholdersForm"
-        (skip)="skip()"
-        skippable
-      ></waterfall-right-holder-form>
-      <section>
-        <ng-container *ngIf="updating$ | async; else editRightHolders">
-          <logo-spinner></logo-spinner>
-        </ng-container>
-        <ng-template #editRightHolders>
-          <button mat-flat-button color="primary" (click)="update()" [disabled]="invalidRightholders$ | async">
-            <span>Continue</span>
-            <mat-icon svgIcon="arrow_forward"></mat-icon>
-          </button>
-        </ng-template>
-      </section>
-    </mat-step>
-    <mat-step>
-      <ng-template matStepLabel>Add Contracts</ng-template>
-      <ng-template matStepContent>
-        <waterfall-contracts-form [movieId]="movieId" [documentForm]="documentForm" (skip)="skip()"></waterfall-contracts-form>
-        <section>
-          <ng-container *ngIf="updating$ | async; else editContracts">
-            <logo-spinner></logo-spinner>
-          </ng-container>
-          <ng-template #editContracts>
-            <a mat-flat-button color="primary" [routerLink]="['..']">
-              <span>Continue</span>
-              <mat-icon svgIcon="arrow_forward"></mat-icon>
-            </a>
-          </ng-template>
-        </section>
-      </ng-template>
-    </mat-step>
-    
-  </mat-stepper>
-
-</ng-template>
